Fix `Domainic::Type::Definition`
This ensures `Domainic::Type::Definition` properly injects types onto
itself when called.  Unfortunately this is still unable to overwrite
ruby core types which will be handled in a separate issue. For now the
workaround is either to use the global prefix or use an alias for the
type that doesn't conflict with a ruby core type.

Changelog:
  - changed `Domainic::Type::Definition` now properly injects types
    onto itself.

see #105