Fixes `Domainic::Dev::CLI::Command::LintMarkdownCommand` (#105)
## Description

This fixes the ordering of mdl violations in the `bin/dev lint md`
output. Additionally, this enhances the output of the violations to link
to the appropriate file and line number; as well as link to the mdl rule
being violated. Simply click the line number or mdl rule id, and you
will be brought to the appropriate place.

## Related Issues

- fixes #97

## Changes Made

- mdl violations are now output in line order
- the line numbers of each mdl violation will go directly to the
offending file and line number when clicked
- the mdl rule ids of each mdl violation will go directly to the
documentation for that rule when clicked.

## Checklist

Before submitting this PR, please ensure:
* [x] I have run `bin/dev ci` and all checks pass
* [x] I have added/updated tests that prove my fix/feature works
* [x] I have added/updated documentation as needed